# apps/backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for uploading PDF files and extracting images from them\n\nThis API provides endpoints for:\n- Uploading PDF files directly or via presigned URLs\n- Processing PDFs to extract images\n- Downloading extracted images as ZIP archives",
        "title": "PDF to Images API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/jobs": {
            "post": {
                "description": "Create a new job and get a presigned URL for uploading a PDF file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Initialize PDF upload",
                "operationId": "initPDFUpload",
                "responses": {
                    "201": {
                        "description": "Upload initialized successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.InitUploadResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to initialize upload",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{job_id}/complete-upload": {
            "post": {
                "description": "Mark the PDF upload as complete and start processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Complete PDF upload",
                "operationId": "completePDFUpload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload completed successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid path or job id",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to complete upload",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pdf": {
            "post": {
                "description": "Upload a PDF file for processing",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF"
                ],
                "summary": "Upload PDF file",
                "operationId": "uploadPDF",
                "parameters": [
                    {
                        "type": "file",
                        "description": "PDF file to upload",
                        "name": "pdf",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF uploaded successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.PDFUploadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or file format",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/zip": {
            "get": {
                "description": "Download all processed images from PDFs as a ZIP archive",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "ZIP"
                ],
                "summary": "Download processed images as ZIP",
                "operationId": "getZip",
                "responses": {
                    "200": {
                        "description": "ZIP file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.InitUploadResponse": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string"
                },
                "pdf_key": {
                    "type": "string"
                },
                "presigned_url": {
                    "type": "string"
                }
            }
        },
        "handler.PDFUploadSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}